fix: rename slack/discord modules to avoid PyPI package collision
app.py did 'from slack import ...' but our slack.py collides with the
installed 'slack' PyPI package. Locally our module wins (api/src is
first on sys.path); on Lambda everything is flattened into /var/task
where the 'slack' *package* dir shadows our module, and it imports
aiohttp (not installed) -> whole app fails to import -> every request
500s. Rename to slack_integration/discord_integration so the import can
only resolve to our code. Reproduced the collision locally and verified
the app loads with a broken 'slack' package shadowing on the path.
